% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometric_euclidean.R
\name{gmed_euclidean}
\alias{gmed_euclidean}
\title{Geometric Median in Euclidean Space}
\usage{
gmed_euclidean(X, weights = NULL, ...)
}
\arguments{
\item{X}{an \eqn{(n\times p)} matrix for \eqn{p}-dimensional signal. If vector is given, it is assumed that \eqn{p=1}.}

\item{weights}{\code{NULL} for equal weight \code{rep(1/n,n)}; otherwise, it has to be a vector of length \eqn{n}.}

\item{...}{extra parameters:\describe{
\item{maxiter}{maximum number of iterations (default: 100).}
\item{abstol}{Cauchy stopping criterion (default: 1e-8).}
}}
}
\value{
a named list containing\describe{
\item{solution}{an optimal solution as a length-\eqn{p} vector.}
\item{time}{elapsed wall-clock time for computation measured in seconds.}
}
}
\description{
This function solves a weighted version of the L1-median problem, which is 
geometric median under the Euclidean setting. For a given data 
\eqn{x_1, x_2, \ldots, x_n \in \mathbb{R}^p}, it minimizes the following functional:
\deqn{F(\mu) = \sum_{i=1}^n w_i \|x_i - \mu\|_2,}
for some nonnegative weights \eqn{w_1, \ldots, w_n \in \mathbb{R}}.
}
\examples{
\donttest{
## set seed for reproducibility
set.seed(496)

## generate sin(x) data with noise for 100 replicates
X = seq(from=0,to=10,length.out=20)
Y = array(0,c(100,20))
for (i in 1:100){
   Y[i,] = sin(X) + stats::rnorm(20, sd=0.75)
}

## compute L1-median and L2-mean
vecL2 = base::colMeans(Y)
vecL1 = gmed_euclidean(Y)$solution

## compute deviation from the signal
signal = sin(X)
devL2  = sqrt(sum(vecL2-signal)^2)
devL1  = sqrt(sum(vecL1-signal)^2)

## visualize
opar <- par(no.readonly=TRUE)
par(mfrow=c(1,3), pty="s")
matplot(t(Y[1:5,]), type="l", main="5 generated data", ylim=c(-2,2))
plot(t, vecL2, type="l", col="blue", main=paste0("L2:error=",round(devL2,4)), ylim=c(-2,2))
plot(t, vecL1, type="l", col="red",  main=paste0("L1:error=",round(devL1,4)), ylim=c(-2,2))
par(opar)
}

}
\concept{geometric}
