% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometric_grassmann.R
\name{gmed_grassmann}
\alias{gmed_grassmann}
\title{Geometric Median on Grassmann Manifold}
\usage{
gmed_grassmann(X, weights = NULL, ...)
}
\arguments{
\item{X}{X an \eqn{(p\times k\times n)} array that stacks \eqn{n} \eqn{k}-frames. It must be provided as a 3-dimensional array.}

\item{weights}{\code{NULL} for equal weight \code{rep(1/n,n)}; otherwise, it has to be a vector of length \eqn{n}.}

\item{...}{extra parameters:\describe{
\item{maxiter}{maximum number of iterations (default: 100).}
\item{abstol}{Cauchy stopping criterion (default: 1e-8).}
}}
}
\value{
a named list containing\describe{
\item{solution}{an optimal solution as an \eqn{(p\times k)} orthonormal basis.}
\item{time}{elapsed wall-clock time for computation measured in seconds.}
}
}
\description{
Grassmann manifold \eqn{Gr(p,k)} is a collection of \eqn{k}-dimensional subspaces in \eqn{\mathbb{R}^p}. 
For a collection of \eqn{k}-frames, it computes a weighted version of the geometric median 
under the chordal distance.
}
\examples{
\donttest{
## load the iris data and extract the data
data("iris")

X = as.matrix(iris[,1:4])
lab = as.factor(iris[,5])

## repeat 80\% sampling and getting 2-dimensional PCA
proj3 <- array(0,c(4,2,50)) # repeat this 50 times
for (it in 1:50){
  # sampling
  Xsam = X[sample(1:nrow(X),round(nrow(X)*0.8)),]
  
  # compute eigenvector of empirical covariance
  Xcov = stats::cov(Xsam)
  Xeig = base::eigen(Xcov)
  
  # same the top-2 PCs
  proj3[,,it] = Xeig$vectors[,1:2]
}

## compute geometric median
grmed = gmed_grassmann(proj3)
proj_gmed = grmed$solution

## compute PCA for the original data
proj_base = base::eigen(stats::cov(X))$vectors[,1:2]

## visualize
opar <- par(no.readonly=TRUE)
par(mfrow=c(1,2), pty="s")
plot(X\%*\%proj_base, col=lab, pch=19, xlab="", ylab="", main="base PCA")
plot(X\%*\%proj_gmed, col=lab, pch=19, xlab="", ylab="", main="geometric median")
par(opar)
}

}
\concept{geometry}
