\name{medians-package}
\alias{medians-package}
\alias{medians}
\docType{package}
\title{\packageTitle{medians}}
\description{\packageDescription{medians}}
\section{Package Content}{\packageIndices{medians}}
\author{\packageAuthor{medians}}
\section{Maintainer}{\packageMaintainer{medians}}
\keyword{package}
